requirements.txt
python-dotenv
alpaca-py
newspaper3k